+++ app/globals.ts
@@ -5,4 +5,4 @@
 
 export var sep='/';
 export var version: string="1.0.0-beta.03";   
-export var themeName='theme1';
+export var themeName='theme2';